fix:compile error

+++ examples/c/tmq.c
@@ -280,7 +280,7 @@ void consume_repeatly(tmq_t* tmq) {
 
     code = tmq_offset_seek(tmq, topic_name, p->vgId, p->begin);
     if (code != 0) {
-      fprintf(stderr, "failed to seek to %lld, reason:%s", p->begin, tmq_err2str(code));
+      fprintf(stderr, "failed to seek to %d, reason:%s", (int)p->begin, tmq_err2str(code));
     }
   }
